ESC-823 initial relocation of code handling business entity self removal from undertaking
